salary integer -> bigDecimal

+++ src/main/java/ru/talismanandco/jobhelpercore/entity/SalaryEntity.java
@@ -4,14 +4,15 @@ import lombok.Data;
 
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
+import java.math.BigDecimal;
 
 @Embeddable
 @Data
 public class SalaryEntity {
     @Column(name = "bottom")
-    private int from;
+    private BigDecimal from;
     @Column(name = "top")
-    private int to;
+    private BigDecimal to;
     private String currency;
     private boolean gross;
 }
